requirements.txt
Django==2.2.19
djangorestframework==3.13.1
